Improve `number-input` stepping behaviour
Stepping now starts empty inputs from the placeholder number, if defined as such.
If the above condition doesn't apply, the first stepped value to appear on an empty input will now be `1` or `-1` (depending of the step direction), clamped to the min/max numbers, instead of starting with the next sibling of  `1` or `-1`, also clamped to the min/max numbers.